requirements_basic.txt
# Dependencias básicas para Barbara
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.30.0
python-dotenv>=1.0.0
twilio>=8.0.0
markdown>=3.4.0
jinja2>=3.0.0

# Dependencias opcionales (comentadas para evitar problemas)
# spacy>=3.8.0
# psycopg2-binary>=2.9.0
# openai>=1.0.0
